Compile to static library
