+++ src/include/converter.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "standard.hpp"
+
 namespace utils
 {
     /**
@@ -8,11 +10,7 @@ namespace utils
      * @param str An `std::string` to convert
      * @return The converted `std::wstring`
      */
-    std::wstring utf_convert(const std::string &str)
-    {
-        std::wstring_convert<std::codecvt_utf8<wchar_t>, wchar_t> converter;
-        return converter.from_bytes(str);
-    }
+    std::wstring utf_convert(const std::string &str);
 
     /**
      * @brief Convert `std::wstring` to `std::wstring`
@@ -20,9 +18,5 @@ namespace utils
      * @param wstr An `std::wstring` to convert
      * @return The converted `std::string`
      */
-    std::string utf_convert(const std::wstring &wstr)
-    {
-        std::wstring_convert<std::codecvt_utf8<wchar_t>, wchar_t> converter;
-        return converter.to_bytes(wstr);
-    }
+    std::string utf_convert(const std::wstring &wstr);
 }
