+++ src/include/error.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "format.hpp"
+#include "standard.hpp"
+
 namespace liteshell
 {
     /** @brief Base class for specific exceptions from this application */
@@ -9,75 +12,71 @@ namespace liteshell
         /** @brief The error message */
         const std::string message;
 
-        LiteShellException(const std::string &message) : message(message) {}
+        LiteShellException(const std::string &message);
 
-        const char *what() const noexcept
-        {
-            return message.c_str();
-        }
+        const char *what() const noexcept;
     };
 
     /** @brief Exceptions regarding subprocesses */
     class SubprocessException : public LiteShellException
     {
     public:
-        SubprocessException(const std::string &message) : LiteShellException(message) {}
+        SubprocessException(const std::string &message);
     };
 
     /** @brief Exception thrown when a subprocess cannot be created */
     class SubprocessCreationError : public SubprocessException
     {
     public:
-        SubprocessCreationError(const std::string &message) : SubprocessException(message) {}
+        SubprocessCreationError(const std::string &message);
     };
 
     /** @brief Exceptions regarding shell environment */
     class EnvironmentException : public LiteShellException
     {
     public:
-        EnvironmentException(const std::string &message) : LiteShellException(message) {}
+        EnvironmentException(const std::string &message);
     };
 
     /** @brief Exception thrown when resolving environment variables failed */
     class EnvironmentResolveError : public EnvironmentException
     {
     public:
-        EnvironmentResolveError(const std::string &message) : EnvironmentException(message) {}
+        EnvironmentResolveError(const std::string &message);
     };
 
     /** @brief Exception thrown when a command couldn't be found */
     class CommandNotFound : public LiteShellException
     {
     public:
-        CommandNotFound(const std::string &name, const std::string &suggestion)
-            : LiteShellException(utils::format("Command \"%s\" not found. Did you mean \"%s\"?", name.c_str(), suggestion.c_str())) {}
+        CommandNotFound(const std::string &name, const std::string &suggestion);
     };
 
     /** @brief Exceptions regarding context parsing */
     class ContextException : public LiteShellException
     {
     public:
-        ContextException(const std::string &message) : LiteShellException(message) {}
+        ContextException(const std::string &message);
     };
 
     /** @brief Exception thrown when a required argument is missing */
     class ArgumentMissingError : public ContextException
     {
     public:
-        ArgumentMissingError(const std::string &name) : ContextException(utils::format("Argument \"%s\" is missing", name.c_str())) {}
+        ArgumentMissingError(const std::string &name);
     };
 
     /** @brief Exception thrown when an unrecognized option is provided */
     class UnrecognizedOption : public ContextException
     {
     public:
-        UnrecognizedOption(const std::string &name) : ContextException(utils::format("Unrecognized option \"%s\"", name.c_str())) {}
+        UnrecognizedOption(const std::string &name);
     };
 
     /** @brief Exception thrown when the number of provided positional arguments exceeds the constraint */
     class TooManyPositionalArguments : public ContextException
     {
     public:
-        TooManyPositionalArguments() : ContextException("Too many positional arguments were passed") {}
+        TooManyPositionalArguments();
     };
 }
