+++ src/include/join.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "standard.hpp"
+
 namespace utils
 {
     /**
@@ -14,22 +16,7 @@ namespace utils
     std::string join(
         const _ForwardIterator &__begin,
         const _ForwardIterator &__end,
-        const std::string &delimiter)
-    {
-        std::string result;
-        for (auto iter = __begin; iter != __end; iter++)
-        {
-            result += *iter;
-
-            auto peek = iter;
-            if (++peek != __end)
-            {
-                result += delimiter;
-            }
-        }
-
-        return result;
-    }
+        const std::string &delimiter);
 
     /**
      * @brief Join 2 Windows paths
@@ -38,26 +25,5 @@ namespace utils
      * @param second The second path
      * @return The joined path
      */
-    std::string join(std::string first, std::string second)
-    {
-        {
-            std::size_t size = first.size();
-            while (size > 0 && first[size - 1] == '\\')
-            {
-                size--;
-            }
-            first = first.substr(0, size);
-        }
-
-        {
-            std::size_t index = 0;
-            while (index + 1 < second.size() && second[index] == '\\')
-            {
-                index++;
-            }
-            second = second.substr(index);
-        }
-
-        return first + '\\' + second;
-    }
+    std::string join(std::string first, std::string second);
 }
