+++ src/include/subprocess.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "utils.hpp"
+
 namespace liteshell
 {
     /**
@@ -22,8 +24,7 @@ namespace liteshell
         const std::string command;
 
         /** @brief Construct a new `ProcessInfoWrapper` object */
-        ProcessInfoWrapper(const PROCESS_INFORMATION &info, const std::string &command)
-            : _info(info), command(command) {}
+        ProcessInfoWrapper(const PROCESS_INFORMATION &info, const std::string &command);
 
         /**
          * @brief Destructor for this object.
@@ -32,107 +33,50 @@ namespace liteshell
          * [`CloseHandle`](https://learn.microsoft.com/en-us/windows/win32/api/handleapi/nf-handleapi-closehandle)
          * to close the underlying handles of the subprocess.
          */
-        ~ProcessInfoWrapper()
-        {
-            CloseHandle(_info.hProcess);
-            CloseHandle(_info.hThread);
-        }
+        ~ProcessInfoWrapper();
 
         /** @brief Whether the subprocess is suspended */
-        bool is_suspended() const
-        {
-            return _suspended;
-        }
+        bool is_suspended() const;
 
         /** @brief Suspend the subprocess */
-        void suspend()
-        {
-            if (!_suspended)
-            {
-                if (SuspendThread(_info.hThread) == (DWORD)-1)
-                {
-                    throw std::runtime_error(utils::last_error("SuspendThread ERROR"));
-                }
-
-                _suspended = true;
-            }
-            else
-            {
-                throw std::runtime_error("The process has already been suspended");
-            }
-        }
+        void suspend();
 
         /** @brief Resume the subprocess */
-        void resume()
-        {
-            if (_suspended)
-            {
-                if (ResumeThread(_info.hThread) == (DWORD)-1)
-                {
-                    throw std::runtime_error(utils::last_error("ResumeThread ERROR"));
-                }
-
-                _suspended = false;
-            }
-            else
-            {
-                throw std::runtime_error("The process hasn't been suspended yet");
-            }
-        }
+        void resume();
 
         /**
          * @brief Wait for the subprocess with timeout
          *
          * @param milliseconds The timeout in milliseconds, may equal to `INFINITE`
          */
-        void wait(DWORD milliseconds)
-        {
-            WaitForSingleObject(_info.hProcess, milliseconds);
-        }
+        void wait(DWORD milliseconds);
 
         /**
          * @brief Kill the subprocess
          *
          * @param exit_code The exit code of the subprocess
          */
-        void kill(UINT exit_code)
-        {
-            if (!TerminateProcess(_info.hProcess, exit_code))
-            {
-                throw std::runtime_error(utils::last_error("TerminateProcess ERROR"));
-            }
-        }
+        void kill(UINT exit_code);
 
         /**
          * @brief Get the exit code of the subprocess
          *
          * @return The exit code of the subprocess
          */
-        DWORD exit_code() const
-        {
-            DWORD exit_code;
-            GetExitCodeProcess(_info.hProcess, &exit_code);
-            return exit_code;
-        }
+        DWORD exit_code() const;
 
         /**
          * @brief Get the process ID of the subprocess
          *
          * @return The process ID of the subprocess
          */
-        DWORD pid() const
-        {
-            return _info.dwProcessId;
-        }
+        DWORD pid() const;
 
         /**
          * @brief Get the thread ID of the subprocess
          *
          * @return The thread ID of the subprocess
          */
-        DWORD tid() const
-        {
-            return _info.dwThreadId;
-        }
+        DWORD tid() const;
     };
 }
