+++ src/include/utils.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "converter.hpp"
+#include "format.hpp"
 #include "join.hpp"
 
 namespace utils
@@ -12,10 +13,7 @@ namespace utils
      * @param message The message to prepend to the error
      * @return The formatted error message
      */
-    std::string last_error(const std::string &message)
-    {
-        return format("%s: %d", message.c_str(), GetLastError());
-    }
+    std::string last_error(const std::string &message);
 
     /**
      * @brief Get the size of the console window using
@@ -23,90 +21,25 @@ namespace utils
      *
      * @return The number of columns and rows, respectively.
      */
-    std::pair<SHORT, SHORT> get_console_size()
-    {
-        CONSOLE_SCREEN_BUFFER_INFO info;
-        if (!GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &info))
-        {
-            throw std::runtime_error(last_error("GetConsoleScreenBufferInfo ERROR"));
-        }
-
-        SHORT columns = info.srWindow.Right - info.srWindow.Left + 1,
-              rows = info.srWindow.Bottom - info.srWindow.Top + 1;
-
-        return std::make_pair(columns, rows);
-    }
+    std::pair<SHORT, SHORT> get_console_size();
 
     /** @brief Get the current working directory */
-    std::string get_working_directory()
-    {
-        wchar_t buffer[MAX_PATH];
-        auto size = GetCurrentDirectoryW(MAX_PATH, buffer);
-        if (size == 0)
-        {
-            throw std::runtime_error(last_error("GetCurrentDirectoryW ERROR"));
-        }
-
-        return utf_convert(std::wstring(buffer, buffer + size));
-    }
+    std::string get_working_directory();
 
     /**
      * @brief Get the path of the executable i.e. the path of shell.exe using
      * [`GetModuleFileNameW`](https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-getmodulefilenamew).
      */
-    std::string get_executable_path()
-    {
-        wchar_t buffer[MAX_PATH];
-        auto size = GetModuleFileNameW(NULL, buffer, MAX_PATH);
-        if (size == 0)
-        {
-            throw std::runtime_error(last_error("Error calling GetModuleFileNameW"));
-        }
-
-        return utf_convert(std::wstring(buffer, buffer + size));
-    }
+    std::string get_executable_path();
 
     /** @brief The CTRL handler used by the command shell */
-    BOOL WINAPI ctrl_handler(DWORD ctrl_type)
-    {
-        return ctrl_type == CTRL_C_EVENT;
-    }
+    BOOL WINAPI ctrl_handler(DWORD ctrl_type);
 
     /** @brief Set whether to ignore Ctrl C signal */
-    void set_ignore_ctrl_c(bool ignore)
-    {
-        if (!SetConsoleCtrlHandler(ctrl_handler, ignore))
-        {
-            std::cerr << last_error("Warning: SetConsoleCtrlHandler ERROR") << std::endl;
-        }
-    }
+    void set_ignore_ctrl_c(bool ignore);
 
     /** @brief List all files matching a specific pattern (typically used to list a directory) */
-    std::vector<WIN32_FIND_DATAW> list_files(const std::string &__pattern)
-    {
-        std::vector<WIN32_FIND_DATAW> results(1);
-
-        HANDLE h_file = FindFirstFileW(utf_convert(__pattern).c_str(), &results[0]);
-        if (h_file == INVALID_HANDLE_VALUE)
-        {
-            return {};
-        }
-
-        do
-        {
-            results.emplace_back();
-        } while (FindNextFileW(h_file, &results.back()));
-
-        results.pop_back();
-
-        if (!FindClose(h_file))
-        {
-            throw std::runtime_error(last_error("Error when closing file search handle"));
-        }
-
-        CloseHandle(h_file);
-        return results;
-    }
+    std::vector<WIN32_FIND_DATAW> list_files(const std::string &__pattern);
 
     /**
      * @brief Remove a directory recursively. It is the caller's responsibility to ensure that the directory exists.
@@ -115,50 +48,7 @@ namespace utils
      * @param verbose Whether to display the deletion progress
      * @return Whether the directory was removed successfully
      */
-    bool remove_directory(const std::string &directory, bool verbose)
-    {
-        bool success = true;
-        for (const auto &file : list_files(join(directory, "*")))
-        {
-            auto filename = utf_convert(file.cFileName);
-
-            // Skip current and parent directory
-            if (filename == "." || filename == "..")
-            {
-                continue;
-            }
-
-            auto path = join(directory, filename);
-            if (file.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY)
-            {
-                if (!remove_directory(path, verbose))
-                {
-                    success = false;
-                }
-            }
-            else if (DeleteFileW(utf_convert(path).c_str()))
-            {
-                std::cout << "Deleted " << path << std::endl;
-            }
-            else
-            {
-                std::cerr << last_error(format("Error deleting file \"%s\"", path.c_str())) << std::endl;
-                success = false;
-            }
-        }
-
-        if (RemoveDirectoryW(utf_convert(directory).c_str()))
-        {
-            std::cout << "Deleted " << directory << std::endl;
-        }
-        else
-        {
-            std::cerr << last_error(format("Error deleting directory \"%s\"", directory.c_str())) << std::endl;
-            success = false;
-        }
-
-        return success;
-    }
+    bool remove_directory(const std::string &directory, bool verbose);
 
     /**
      * @brief Whether a string has the specified prefix
@@ -167,23 +57,7 @@ namespace utils
      * @param prefix The prefix to check for
      * @return Whether the string has the specified prefix
      */
-    bool startswith(const std::string &string, const std::string &prefix)
-    {
-        if (string.size() < prefix.size())
-        {
-            return false;
-        }
-
-        for (std::size_t i = 0; i < prefix.size(); i++)
-        {
-            if (string[i] != prefix[i])
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool startswith(const std::string &string, const std::string &prefix);
 
     /**
      * @brief Whether a string has the specified suffix
@@ -192,23 +66,7 @@ namespace utils
      * @param suffix The suffix to check for
      * @return Whether the string has the specified suffix
      */
-    bool endswith(const std::string &string, const std::string &suffix)
-    {
-        if (string.size() < suffix.size())
-        {
-            return false;
-        }
-
-        for (std::size_t i = string.size() - suffix.size(); i < string.size(); i++)
-        {
-            if (string[i] != suffix[i + suffix.size() - string.size()])
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    bool endswith(const std::string &string, const std::string &suffix);
 
     /**
      * @brief Escape a string to obtain a valid regex pattern
@@ -216,20 +74,7 @@ namespace utils
      * @param original The original string
      * @return The escaped string
      */
-    std::string regex_escape(const std::string &original)
-    {
-        std::stringstream result;
-        for (char c : original)
-        {
-            if (c == '.' || c == '*' || c == '+' || c == '?' || c == '|' || c == '(' || c == ')' ||
-                c == '[' || c == ']' || c == '{' || c == '}' || c == '\\' || c == '^' || c == '$')
-            {
-                result << '\\';
-            }
-            result << c;
-        }
-        return result.str();
-    }
+    std::string regex_escape(const std::string &original);
 
     /**
      * @brief Whether a file is executable
@@ -237,19 +82,12 @@ namespace utils
      * @param name The name of the file
      * @return Whether the file is executable
      */
-    bool is_executable(LPCWSTR name)
-    {
-        DWORD _;
-        return GetBinaryTypeW(name, &_);
-    }
+    bool is_executable(LPCWSTR name);
 
     /**
      * @brief A helper method equivalent to `predicate ? first : second`
      */
-    std::string ngettext(const bool predicate, const std::string &first, const std::string &second)
-    {
-        return predicate ? first : second;
-    }
+    std::string ngettext(const bool predicate, const std::string &first, const std::string &second);
 
     /**
      * @brief Whether a character is a mathematical symbol
@@ -257,10 +95,7 @@ namespace utils
      * @param c The character to check
      * @return Whether the character is a mathematical symbol
      */
-    bool is_math_symbol(char c)
-    {
-        return c == '+' || c == '-' || c == '*' || c == '/' || c == ' ' || c == '%' || ('0' <= c && c <= '9') || c == '(' || c == ')';
-    }
+    bool is_math_symbol(char c);
 
     const boost::regex _command_name = boost::regex(R"(^\w+$)");
 
@@ -270,10 +105,7 @@ namespace utils
      * @param name The name to check
      * @return Whether the name is a valid command name
      */
-    bool is_valid_command(const std::string &name)
-    {
-        return boost::regex_match(name, _command_name);
-    }
+    bool is_valid_command(const std::string &name);
 
     const boost::regex _variable_name = boost::regex(R"(^\w+$)");
 
@@ -283,10 +115,7 @@ namespace utils
      * @param name The name to check
      * @return Whether the name is a valid variable name
      */
-    bool is_valid_variable(const std::string &name)
-    {
-        return boost::regex_match(name, _variable_name);
-    }
+    bool is_valid_variable(const std::string &name);
 
     const boost::regex _short_option_name = boost::regex(R"(^-[a-zA-Z]$)");
 
@@ -296,10 +125,7 @@ namespace utils
      * @param name The name to check
      * @return Whether the name is a valid short option name
      */
-    bool is_valid_short_option(const std::string &name)
-    {
-        return boost::regex_match(name, _short_option_name);
-    }
+    bool is_valid_short_option(const std::string &name);
 
     const boost::regex _long_option_name = boost::regex(R"(^--[a-zA-Z\-_]+$)");
 
@@ -309,10 +135,7 @@ namespace utils
      * @param name The name to check
      * @return Whether the name is a valid long option name
      */
-    bool is_valid_long_option(const std::string &name)
-    {
-        return boost::regex_match(name, _long_option_name);
-    }
+    bool is_valid_long_option(const std::string &name);
 
     /**
      * @brief Get the status of global memory using the native method
@@ -320,17 +143,7 @@ namespace utils
      *
      * @return The status of global memory
      */
-    MEMORYSTATUSEX get_global_memory_status()
-    {
-        MEMORYSTATUSEX memory_status;
-        memory_status.dwLength = sizeof(memory_status);
-        if (!GlobalMemoryStatusEx(&memory_status))
-        {
-            throw std::runtime_error(last_error("GlobalMemoryStatusEx ERROR"));
-        }
-
-        return memory_status;
-    }
+    MEMORYSTATUSEX get_global_memory_status();
 
     /**
      * @brief Calculate the square root of a value using binary search
@@ -339,39 +152,7 @@ namespace utils
      * @return The square root of the value
      */
     template <typename T>
-    T sqrt(const T &value)
-    {
-        if (value < 0)
-        {
-            throw std::out_of_range(format("Attempted to calculate square root of %s < 0", std::to_string(value).c_str()));
-        }
-
-        if (value == 0)
-        {
-            return 0;
-        }
-
-        T low = 0, high = std::max(static_cast<T>(1), value), accuracy = 1;
-        if constexpr (std::is_floating_point_v<T>)
-        {
-            accuracy = 1.0e-6;
-        }
-
-        while (high - low > accuracy)
-        {
-            double mid = (low + high) / 2;
-            if (mid * mid < value)
-            {
-                low = mid;
-            }
-            else
-            {
-                high = mid;
-            }
-        }
-
-        return high;
-    }
+    T sqrt(const T &value);
 
     /**
      * @brief Register a callback to run when this object is destroyed
@@ -386,12 +167,9 @@ namespace utils
         Finalize &operator=(const Finalize &) = delete;
 
     public:
-        /** Construct a new `Finalize` object with a registered callback */
-        Finalize(const std::function<void()> &callback) : callback(callback) {}
+        /** @brief Construct a new `Finalize` object with a registered callback */
+        Finalize(const std::function<void()> &callback);
 
-        ~Finalize()
-        {
-            callback();
-        }
+        ~Finalize();
     };
 }
